+++ proto-tcp/client.go
@@ -10,7 +10,7 @@ const (
 	_DIAL_TIMEOUT = 10 * time.Second
 )
 
-func NewClient(host string, port uint16) (*Conn, error) {
+func NewClient(host string, port int) (*Conn, error) {
 	nconn, err := net.DialTimeout("tcp", fmt.Sprintf("%s:%d", host, port), _DIAL_TIMEOUT)
 	if err != nil {
 		return nil, err
